mcp/requirements.txt
mcp[http,fastapi]>=1.10.1
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
google-adk>=0.1.0
Pillow>=9.0.0
